backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
celery==5.3.4
redis==5.0.1
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
pydantic==2.5.0
python-multipart==0.0.6
openai==1.3.8
tweepy==4.14.0
praw==7.7.1
requests==2.31.0
python-dotenv==1.0.0
websockets==12.0